# docs/swagger.yaml
basePath: /
definitions:
  main.Exchange:
    properties:
      created_at:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      requester_id:
        type: integer
      service_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  main.ExchangeRequest:
    properties:
      owner_id:
        type: integer
      requester_id:
        type: integer
      service_id:
        type: integer
    type: object
  main.Review:
    properties:
      author_id:
        type: integer
      commentaire:
        type: string
      created_at:
        type: string
      exchange_id:
        type: integer
      id:
        type: integer
      note:
        type: integer
      target_id:
        type: integer
    type: object
  main.ReviewRequest:
    properties:
      commentaire:
        type: string
      note:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST d'échange de compétences via crédits-temps.
  title: BarterSwap API
  version: "1.0"
paths:
  /api/exchanges:
    get:
      description: Récupère la liste des demandes d'échanges (en tant que demandeur
        ou prestataire) pour l'utilisateur connecté, avec filtre par statut optionnel.
      parameters:
      - description: ID de l'utilisateur connecté
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Filtrer par statut (pending, accepted, rejected, cancelled, completed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Exchange'
            type: array
        "400":
          description: En-tête manquant
          schema:
            type: string
      summary: Lister mes échanges
      tags:
      - Exchanges
    post:
      consumes:
      - application/json
      description: Crée une nouvelle demande d'échange pour un service spécifié. L'utilisateur
        connecté doit disposer d'assez de crédits.
      parameters:
      - description: ID de l'utilisateur connecté
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Détails de la demande d'échange
        in: body
        name: exchange
        required: true
        schema:
          $ref: '#/definitions/main.ExchangeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: Requête invalide ou crédits insuffisants
          schema:
            type: string
        "404":
          description: Service ou demandeur non trouvé
          schema:
            type: string
        "409":
          description: Conflit (échange déjà actif)
          schema:
            type: string
      summary: Créer une demande d'échange
      tags:
      - Exchanges
  /api/exchanges/{id}:
    get:
      description: Récupère les détails d'un échange par son ID.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
      summary: Obtenir le détail d'un échange
      tags:
      - Exchanges
  /api/exchanges/{id}/accept:
    put:
      description: Accepte une demande d'échange en cours. Seul le propriétaire du
        service peut accepter et cela débite les crédits correspondants au demandeur.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      - description: ID du propriétaire
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: Requête invalide ou crédits insuffisants
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
      summary: Accepter une demande d'échange
      tags:
      - Exchanges
  /api/exchanges/{id}/cancel:
    put:
      description: Annule un échange en cours. Le demandeur et le prestataire peuvent
        tous deux l'annuler. Si l'échange était accepté, le demandeur est remboursé.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      - description: ID de l'utilisateur connecté
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: Statut invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
      summary: Annuler un échange
      tags:
      - Exchanges
  /api/exchanges/{id}/complete:
    put:
      description: Termine un échange. Seul le demandeur peut le valider une fois
        le service rendu, ce qui transfère les crédits au prestataire.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      - description: ID du demandeur
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: Statut invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
      summary: Marquer un échange comme terminé
      tags:
      - Exchanges
  /api/exchanges/{id}/reject:
    put:
      description: Refuse une demande d'échange en cours. Seul le propriétaire du
        service peut la refuser.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      - description: ID du propriétaire
        in: header
        name: X-User-ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Exchange'
        "400":
          description: Statut invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
      summary: Refuser une demande d'échange
      tags:
      - Exchanges
  /api/exchanges/{id}/review:
    post:
      consumes:
      - application/json
      description: Enregistre un avis (note et commentaire) pour un échange terminé.
        L'auteur doit être le demandeur ou le prestataire de l'échange.
      parameters:
      - description: ID de l'échange
        in: path
        name: id
        required: true
        type: integer
      - description: ID de l'auteur
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Détails de l'avis
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/main.ReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Review'
        "400":
          description: Requête invalide ou échange non terminé
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Échange non trouvé
          schema:
            type: string
        "409":
          description: Avis déjà soumis
          schema:
            type: string
      summary: Laisser un avis
      tags:
      - Reviews
...
  /api/services/{id}/reviews:
    get:
      description: Récupère la liste de tous les avis laissés sur un service spécifique
        par son ID.
      parameters:
      - description: ID du service
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Review'
            type: array
        "400":
          description: ID invalide
          schema:
            type: string
      summary: Obtenir les avis sur un service
      tags:
      - Reviews
...
  /api/users/{id}/reviews:
    get:
      description: Récupère la liste de tous les avis laissés sur le profil d'un utilisateur
        par son ID.
      parameters:
      - description: ID de l'utilisateur cible
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Review'
            type: array
        "400":
          description: ID invalide
          schema:
            type: string
      summary: Obtenir les avis reçus par un utilisateur
      tags:
      - Reviews
...
swagger: "2.0"